Play wave 2 of the 2026-07-02 revision-ledger sweep
CORRECTNESS: p-adic-guard-gap (WittVec/Qq/Ramified assert_supported_params
+ rejection tests, Qq checked valuation arithmetic incl. a checked_neg Qp
lacks — routed to wave 3), bw-ff-sweep (all-eight-residue ff correction
sweep + the two doc fixes), char2-decomp-coverage (isometric_finite_char2
callers, radical_anisotropic caveat construction, the F4 two-quadratics
input verified to fire the equal-degree trace splitter), spinor-norm-
char2-claim (doc rewrite with the Q(v)^2 counterexample + Nimber-backed
spinor_norm/classify_versor pin), milnor p=3 kappa-twist cross-check,
teichmuller F>1 pin, adele_prec cap verified and widened to the honest
LocalQp bound, AJ "source-pinned" wording reserved for external data pins.

CONSISTENCY: helper-commons (valuation quartet + rdiv + rational_mod_int
into diagonal.rs, linalg::integer::prime_factors, one binomial_checked;
unit_sign_odd vs odd_unit_residue found to be different questions — core
consolidated, both names kept), partiality (isotropy_over_adeles None-
threaded, misere/octal quotient builders return Option<Quotient>),
encapsulation (WittClass, SpinorRep/LazySpinorRep, Cga, LinearMap.cols,
Zp/WittVec reduce-in-neg), display-mirrors (FunctionFieldBrauerWallClass,
WittClassG), engine-cohesion (pure_scalar_norm gate shared, hopf uses
add_term, //! headers on all seven engine files), strip_factor width
unified, ...Isotropy blessed in the forms glossary, Dawson 0.137 cited,
Quotient.num_classes derived not stored.

Lead fixups: py bindings adapted to the new encapsulations (forms.rs
WittClass, engine.rs SpinorRep into_parts/Cga/LinearMap), octal_hunt +
tour examples, private intra-doc links demoted in local_global.

+++ src/forms/springer/char2/asnf.rs
@@ -64,7 +64,7 @@ pub(super) fn valuation<S: FiniteChar2Field>(
         FunctionFieldPlace::Finite(p) => {
             let (mn, _) = strip_factor(a.num().clone(), p);
             let (md, _) = strip_factor(a.den().clone(), p);
-            Some(mn as i128 - md as i128)
+            Some(mn - md)
         }
         FunctionFieldPlace::Infinite => Some(
             a.den().degree().expect("nonzero den") as i128
@@ -88,7 +88,7 @@ fn laurent_finite<S: FiniteChar2Field>(
     }
     let (mn, ncof) = strip_factor(num.clone(), p);
     let (md, e) = strip_factor(den.clone(), p);
-    let val = mn as i128 - md as i128;
+    let val = mn - md;
     let hi_i = n_hi - val; // need power-series digits g_0 .. g_{hi_i}
     if hi_i < 0 {
         return vec![Poly::zero(); len];
